feat: GA 기초 세팅
