+++ apps/admin/app/layout.tsx
@@ -6,6 +6,8 @@ import "@uket/ui/globals.css";
 import { notoSans } from "../config/fonts";
 import Providers from "./providers";
 
+import GoogleAnalytics from "@uket/util/google-analytics";
+
 export const metadata: Metadata = {
   title: "Uket for admin",
   description: "Uket 어드민을 위한 서비스입니다.",
@@ -27,6 +29,7 @@ export default async function RootLayout({
       <body className={`${notoSans.className}`}>
         <Providers>{children}</Providers>
         <Sonner richColors position="bottom-center" />
+        <GoogleAnalytics/>
       </body>
     </html>
   );
