--- apps/web/config/google-analytics.tsx
@@ -1,24 +1,0 @@
-import Script from "next/script";
-
-export default function GoogleAnalytics() {
-  const gtag = process.env.NEXT_PUBLIC_GOOGLE_ANALYTICS_ID;
-
-  return (
-    <>
-      <Script
-        async
-        src={`https://www.googletagmanager.com/gtag/js?id=${gtag}`}
-      />
-      <Script
-        id="google-analytics"
-        dangerouslySetInnerHTML={{
-          __html: `window.dataLayer = window.dataLayer || [];
-  function gtag(){dataLayer.push(arguments);}
-  gtag('js', new Date());
-
-  gtag('config', '${gtag}');`,
-        }}
-      />
-    </>
-  );
-}
