+++ packages/util/package.json
@@ -39,6 +39,7 @@
     "./token": "./src/token.ts",
     "./cookie-client": "./src/cookie-client.ts",
     "./cookie-server": "./src/cookie-server.ts",
-    "./validate-form": "./src/validate-form.ts"
+    "./validate-form": "./src/validate-form.ts",
+    "./google-analytics": "./src/google-analytics.tsx"
   }
 }
